refactor: shared
Story: [ch100]
