--- packages/server/_shared/src/UseCaseError.ts
@@ -1,11 +1,0 @@
-interface IUseCaseError {
-  message: string;
-}
-
-export abstract class UseCaseError implements IUseCaseError {
-  public readonly message: string;
-
-  constructor(message: string) {
-    this.message = message;
-  }
-}
