+++ packages/server/_shared/src/index.ts
@@ -1,7 +1,3 @@
-export { EntityID } from './EntityID';
-export { Entity } from './Entity';
-export { UnexpectedError } from './AppError';
-export { UseCaseError } from './UseCaseError';
-export { Result } from './Result';
+export * from './entity-id';
+export * from './app-error';
 export * from './either';
-export * from './value-object';
