--- packages/server/_shared/src/value-object.ts
@@ -1,7 +1,0 @@
-export class ValueObject<T> {
-  public readonly props: T;
-
-  constructor(props?: T) {
-    this.props = props;
-  }
-}
